requirements.txt
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.2
